Refactored types, improved createBarber mutation
1) In some cases, what is passed to the GraphQL API is
not what is passed to the MongodbAPI. For example, when creating a
barber, there is no profileImageURL in the GraphQL request params,
as it is recieved from the Auth0 profile creation. So, I have split
up the types for the GraphQL API and the internal MongodbAPI. In most
cases types are the same, but this separation might prove useful later

2) Now createBarber mutation is not called by the Auth0 hook. In fact,
it is itself now calling the Auth0 Management API, creating a user there
and assigning a Barber role to him. Only then it creates the barber in
the MongodbAPI